🚧 W0ZCVH task: split context init tests
