+++ .agentplane/tasks/202605190830-W0ZCVH/pr/github-body.md
@@ -19,8 +19,8 @@ Add a TTY dialog for user-run agentplane context init with basic mode informatio
 - Note:
 
 ```text
-Re-verified current PR head 4118cece7b9f after commit amend and PR publication. Focused checks
-remain the same passing set recorded in the task details.
+Adjusted test placement to avoid expanding the oversized init test file; context init behavior and
+changed-file lint remain verified.
 ```
 - Canonical workflow state lives in the task README.
