+++ .agentplane/tasks/202605190830-W0ZCVH/pr/meta.json
@@ -3,7 +3,7 @@
   "branch": "task/202605190830-W0ZCVH/interactive-context-init",
   "created_at": "2026-05-19T08:32:08.899Z",
   "head_sha": "4118cece7b9f88af436ce1c8c56209d1de218cb8",
-  "last_verified_at": "2026-05-19T08:42:47.740Z",
+  "last_verified_at": "2026-05-19T08:46:53.014Z",
   "last_verified_sha": "4118cece7b9f88af436ce1c8c56209d1de218cb8",
   "pr_number": 3924,
   "pr_url": "https://github.com/basilisk-labs/agentplane/pull/3924",
